renamed project and changed imports in modules
