+++ fastapi_application/app/core/__init__.py
