+++ fastapi_application/app/db/db_utils.py
@@ -3,13 +3,12 @@
 from collections.abc import Sequence
 from typing import Literal
 
+from fastapi_app.app.models.base import Base
 from loguru import logger
 import pandas as pd
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm import Session
 from sqlalchemy.sql.selectable import Select
-
-from ricardo.ee.fastapi_app.app.models.base import Base
 
 
 def query_table_as_dataframe(db_session: Session, orm_query: Select) -> pd.DataFrame:
