+++ fastapi_application/app/main.py
@@ -3,12 +3,11 @@
 import sys
 
 from fastapi import FastAPI
+from fastapi_app.app import api_utils
+from fastapi_app.app.api.routers.routers import api_router
+from fastapi_app.app.models.base import init_db
 from loguru import logger
 from starlette.middleware.cors import CORSMiddleware
-
-from ricardo.ee.fastapi_app.app import api_utils
-from ricardo.ee.fastapi_app.app.api.routers.routers import api_router
-from ricardo.ee.fastapi_app.app.models.base import init_db
 
 
 def init_app() -> FastAPI:
@@ -58,4 +57,4 @@ def liveness() -> dict:
     return {"status": "okay"}
 
 
-# uvicorn ricardo.ee.fastapi_app.app.main:app --reload
+# uvicorn fastapi_application.app.main:app --reload
