+++ fastapi_application/app/models/dummy_orm.py
@@ -1,9 +1,8 @@
 """ORM Model."""
 
+from fastapi_app.app.models.base import Base
 from sqlalchemy import Column, DateTime, Integer
 from sqlalchemy.sql import func
-
-from ricardo.ee.fastapi_app.app.models.base import Base
 
 
 class Dummy(Base):
