+++ fastapi_application/app/models/parks_facilities.py
@@ -1,10 +1,9 @@
 """ORM Model."""
 
+from fastapi_app.app.models.base import Base
+from fastapi_app.app.models.parks import Parks
 from sqlalchemy import Column, ForeignKey, Integer, String, Unicode
 from sqlalchemy.orm import relationship
-
-from ricardo.ee.fastapi_app.app.models.base import Base
-from ricardo.ee.fastapi_app.app.models.parks import Parks
 
 
 class ParkFacilities(Base):
