+++ fastapi_application/app/models/species.py
@@ -1,9 +1,8 @@
 """ORM Model."""
 
+from fastapi_app.app.models.base import Base
 from sqlalchemy import Column, Integer, String, Unicode
 from sqlalchemy.orm import relationship
-
-from ricardo.ee.fastapi_app.app.models.base import Base
 
 
 class Species(Base):
