+++ fastapi_application/app/models/visitors.py
@@ -1,9 +1,8 @@
 """ORM Model."""
 
+from fastapi_app.app.models.base import Base
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
-
-from ricardo.ee.fastapi_app.app.models.base import Base
 
 
 class Visitors(Base):
