+++ fastapi_application/app/tests/test_core.py
