+++ fastapi_application/tests/test_core.py
